backend/requirements.txt
# Django Backend Dependencies
django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.3.1
pandas==2.3.3
reportlab==4.4.4
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
